+++ dohproxy/utils.py
@@ -6,7 +6,11 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 #
+import argparse
 import base64
+import urllib.parse
+
+from dohproxy import constants
 
 
 def doh_b64_encode(s):
@@ -24,3 +28,63 @@ def doh_b64_decode(s):
     """
     padding = '=' * (-len(s) % 4)
     return base64.urlsafe_b64decode(s + padding)
+
+
+def build_query_params(dns_query):
+    """Given a wire-format DNS query, build the query parameters.
+    """
+    contenttype = constants.DOH_MEDIA_TYPE
+    return {
+        constants.DOH_BODY_PARAM: doh_b64_encode(dns_query),
+        constants.DOH_CONTENT_TYPE_PARAM: contenttype,
+    }
+
+
+def make_url(domain, uri):
+    """Utility function to return a URL ready to use from a browser or cURL....
+    """
+    p = urllib.parse.ParseResult(
+        scheme='https',
+        netloc=domain,
+        path=uri,
+        params='', query='', fragment='',
+    )
+    return urllib.parse.urlunparse(p)
+
+
+def client_parser_base():
+    """Build a ArgumentParser object with all the default arguments that are
+    useful to both client and stub.
+    """
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        '--domain',
+        default='localhost',
+        help='Domain to make DOH request against. Default: [%(default)s]'
+    )
+    parser.add_argument(
+        '--uri',
+        default=constants.DOH_URI,
+        help='DNS API URI. Default [%(default)s]',
+    )
+    parser.add_argument(
+        '--port',
+        default=443,
+        help='Port to connect to. Default: [%(default)s]'
+    )
+    parser.add_argument(
+        '--post',
+        action='store_true',
+        help='Use HTTP POST instead of GET.'
+    )
+    parser.add_argument(
+        '--debug',
+        action='store_true',
+        help='Prints some debugging output',
+    )
+    parser.add_argument(
+        '--insecure',
+        action='store_true',
+        help=argparse.SUPPRESS,
+    )
+    return parser