+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- code refactor between stub and client
+
 
 ## 0.0.1 - 2018-01-11
 ### Added